Rainbow.Translate

Documentation

single :: Char -> [ByteString] -> [ByteString]

escape :: [ByteString] -> [ByteString]

csi :: [ByteString] -> [ByteString]

sgr :: ([ByteString] -> [ByteString]) -> [ByteString] -> [ByteString]

params :: Show a => [a] -> [ByteString] -> [ByteString]

sgrSingle :: Word -> [ByteString] -> [ByteString]

sgrDouble :: Word -> Word -> [ByteString] -> [ByteString]

normalDefault :: [ByteString] -> [ByteString]

bold :: [ByteString] -> [ByteString]

faint :: [ByteString] -> [ByteString]

italic :: [ByteString] -> [ByteString]

underline :: [ByteString] -> [ByteString]

blink :: [ByteString] -> [ByteString]

inverse :: [ByteString] -> [ByteString]

invisible :: [ByteString] -> [ByteString]

strikeout :: [ByteString] -> [ByteString]

foreBlack :: [ByteString] -> [ByteString]

foreRed :: [ByteString] -> [ByteString]

foreGreen :: [ByteString] -> [ByteString]

foreYellow :: [ByteString] -> [ByteString]

foreBlue :: [ByteString] -> [ByteString]

foreMagenta :: [ByteString] -> [ByteString]

foreCyan :: [ByteString] -> [ByteString]

foreWhite :: [ByteString] -> [ByteString]

foreDefault :: [ByteString] -> [ByteString]

backBlack :: [ByteString] -> [ByteString]

backRed :: [ByteString] -> [ByteString]

backGreen :: [ByteString] -> [ByteString]

backYellow :: [ByteString] -> [ByteString]

backBlue :: [ByteString] -> [ByteString]

backMagenta :: [ByteString] -> [ByteString]

backCyan :: [ByteString] -> [ByteString]

backWhite :: [ByteString] -> [ByteString]

backDefault :: [ByteString] -> [ByteString]

fore256 :: Word8 -> [ByteString] -> [ByteString]

back256 :: Word8 -> [ByteString] -> [ByteString]

foreColor8 :: Color8 -> [ByteString] -> [ByteString]

backColor8 :: Color8 -> [ByteString] -> [ByteString]

foreColor256 :: Color256 -> [ByteString] -> [ByteString]

backColor256 :: Color256 -> [ByteString] -> [ByteString]

styleCommon :: StyleCommon -> [ByteString] -> [ByteString]

style8 :: Style8 -> [ByteString] -> [ByteString]

style256 :: Style256 -> [ByteString] -> [ByteString]

textSpec8 :: TextSpec -> [ByteString] -> [ByteString]

textSpec256 :: TextSpec -> [ByteString] -> [ByteString]

toByteStringsColors0 :: Chunk -> [ByteString] -> [ByteString]

Convert a Chunk to a list of ByteString; do not show any colors. When applied to a Chunk, this function returns a difference list.

toByteStringsColors8 :: Chunk -> [ByteString] -> [ByteString]

Convert a Chunk to a list of ByteString; show eight colors. When applied to a Chunk, this function returns a difference list.

toByteStringsColors256 :: Chunk -> [ByteString] -> [ByteString]

Convert a Chunk to a list of ByteString; show 256 colors. When applied to a Chunk, this function returns a difference list.

byteStringMakerFromEnvironment :: IO (Chunk -> [ByteString] -> [ByteString])

Spawns a subprocess to read the output of tput colors. If this says there are at least 256 colors are available, returns toByteStringsColors256. Otherwise, if there are at least 8 colors available, returns toByteStringsColors8. Otherwise, returns toByteStringsColors0.

If any IO exceptions arise during this process, they are discarded and toByteStringsColors0 is returned.

byteStringMakerFromHandle :: Handle -> IO (Chunk -> [ByteString] -> [ByteString])

Like byteStringMakerFromEnvironment but also consults a provided Handle. If the Handle is not a terminal, toByteStringsColors0 is returned. Otherwise, the value of byteStringMakerFromEnvironment is returned.

chunksToByteStrings

Arguments

:: (Chunk -> [ByteString] -> [ByteString])
Function that converts Chunk to ByteString. This function, when applied to a Chunk, returns a difference list.

-> [Chunk]

-> [ByteString]

Convert a list of Chunk to a list of ByteString. The length of the returned list may be longer than the length of the input list.

So, for example, to print a bunch of chunks to standard output using 256 colors:

{-# LANGUAGE OverloadedStrings #-}
module PrintMyChunks where

import qualified Data.ByteString as BS
import Rainbow

myChunks :: [Chunk]
myChunks = [ "Roses" <> fore red, "\n", "Violets" <> fore blue, "\n" ]

myPrintedChunks :: IO ()
myPrintedChunks = mapM_ BS.putStr
                . chunksToByteStrings toByteStringsColors256
                $ myChunks

To use the highest number of colors that this terminal supports:

myPrintedChunks' :: IO ()
myPrintedChunks' = do
  printer <- byteStringMakerFromEnvironment
  mapM_ BS.putStr
    . chunksToByteStrings printer
    $ myChunks

putChunk :: Chunk -> IO ()

Writes a Chunk to standard output. Spawns a child process to read the output of tput colors to determine how many colors to use, for every single chunk. Therefore, this is not going to win any speed awards. You are better off using chunksToByteStrings and the functions in Data.ByteString to print your Chunks if you are printing a lot of them.

putChunkLn :: Chunk -> IO ()

Writes a Chunk to standard output, and appends a newline. Spawns a child process to read the output of tput colors to determine how many colors to use, for every single chunk. Therefore, this is not going to win any speed awards. You are better off using chunksToByteStrings and the functions in Data.ByteString to print your Chunks if you are printing a lot of them.