Rainbow.Types

Description

The innards of Rainbow. Ordinarily you should not need this module; instead, just import Rainbow, which re-exports the most useful names from this module.

Documentation

type Background8 = Last Color8

type Background256 = Last Color256

type Foreground8 = Last Color8

type Foreground256 = Last Color256

data Enum8

A simple enumeration for eight values.

Constructors

E0

E1

E2

E3

E4

E5

E6

E7

enum8toWord8 :: Enum8 -> Word8

newtype Color8

Color for an 8-color terminal. Does not affect 256-color terminals.

Constructors

Color8

Fields

unColor8 :: Maybe Enum8

Nothing indicates to use the default color for the terminal; otherwise, use the corresponding Terminfo Color.

newtype Color256

Color for an 256-color terminal. Does not affect 8-color terminals.

Constructors

Color256

Fields

unColor256 :: Maybe Word8

Nothing indicates to use the default color for the terminal; otherwise, use the corresponding Terminfo Color.

to256 :: Color8 -> Color256

Any color for an 8-color terminal can also be used in a 256-color terminal.

data StyleCommon

Style elements that apply in both 8 and 256 color terminals. However, the elements are described separately for 8 and 256 color terminals, so that the text appearance can change depending on how many colors a terminal has.

Constructors

StyleCommon

Fields

scBold :: Last Bool

scFaint :: Last Bool

scItalic :: Last Bool

scUnderline :: Last Bool

scBlink :: Last Bool

scInverse :: Last Bool

scInvisible :: Last Bool

scStrikeout :: Last Bool

data Style8

Describes text appearance (foreground and background colors, as well as other attributes such as bold) for an 8 color terminal.

Constructors

Style8

Fields

foreground8 :: Foreground8

background8 :: Background8

common8 :: StyleCommon

data Style256

Describes text appearance (foreground and background colors, as well as other attributes such as bold) for a 256 color terminal.

Constructors

Style256

Fields

foreground256 :: Foreground256

background256 :: Background256

common256 :: StyleCommon

data TextSpec

The TextSpec bundles together the styles for the 8 and 256 color terminals, so that the text can be portrayed on any terminal.

Constructors

TextSpec

Fields

style8 :: Style8

style256 :: Style256

data Chunk

A chunk is some textual data coupled with a description of what color the text is, attributes like whether it is bold or underlined, etc. The chunk knows what foreground and background colors and what attributes to use for both an 8 color terminal and a 256 color terminal.

Constructors

Chunk

Fields

chunkTextSpec :: TextSpec

Specifies all the effects (such as bold, underlining, colors, etc) that apply to this chunk, with different effects for 8 and 256 color terminals; and

chunkTexts :: [Text]

chunkFromText :: Text -> Chunk

Creates a Chunk from a strict Text with default colors and no special effects.

chunkFromTexts :: [Text] -> Chunk

Creates a Chunk from a list of strict Text with default colors and no special effects.

chunkFromLazyText :: Text -> Chunk

Creates a Chunk from a lazy Text with default colors and no special effects.

chunkFromLazyTexts :: [Text] -> Chunk

Creates a Chunk from a list of lazy Text with default colors and no special effects.

bold8 :: Chunk

bold8off :: Chunk

faint8 :: Chunk

faint8off :: Chunk

italic8 :: Chunk

italic8off :: Chunk

underline8 :: Chunk

underline8off :: Chunk

blink8 :: Chunk

blink8off :: Chunk

inverse8 :: Chunk

inverse8off :: Chunk

invisible8 :: Chunk

invisible8off :: Chunk

strikeout8 :: Chunk

strikeout8off :: Chunk

bold256 :: Chunk

bold256off :: Chunk

faint256 :: Chunk

faint256off :: Chunk

italic256 :: Chunk

italic256off :: Chunk

underline256 :: Chunk

underline256off :: Chunk

blink256 :: Chunk

blink256off :: Chunk

inverse256 :: Chunk

inverse256off :: Chunk

invisible256 :: Chunk

invisible256off :: Chunk

strikeout256 :: Chunk

strikeout256off :: Chunk

bold :: Chunk

Bold. What actually happens when you use Bold is going to depend on your terminal. For example, xterm allows you actually use a bold font for bold, if you have one. Otherwise, it might simulate bold by using overstriking. Another possibility is that your terminal might use a different color to indicate bold. For more details (at least for xterm), look at xterm (1) and search for boldColors.

If your terminal uses a different color for bold, this allows an 8-color terminal to really have 16 colors.

boldOff :: Chunk

faint :: Chunk

faintOff :: Chunk

italic :: Chunk

italicOff :: Chunk

underline :: Chunk

underlineOff :: Chunk

blink :: Chunk

blinkOff :: Chunk

inverse :: Chunk

inverseOff :: Chunk

invisible :: Chunk

invisibleOff :: Chunk

strikeout :: Chunk

strikeoutOff :: Chunk